+++ pumpkin-world/src/data/mod.rs
@@ -1,0 +1,1 @@
+pub mod player_data;
