+++ pumpkin-world/src/lib.rs
@@ -5,6 +5,7 @@ pub mod block;
 pub mod chunk;
 pub mod coordinates;
 pub mod cylindrical_chunk_iterator;
+pub mod data;
 pub mod dimension;
 mod generation;
 pub mod item;
