+++ pumpkin/src/data/mod.rs
@@ -9,7 +9,7 @@ pub mod op_data;
 pub mod banlist_serializer;
 pub mod banned_ip_data;
 pub mod banned_player_data;
-pub mod player_data;
+pub mod player_server_data;
 
 pub trait LoadJSONConfiguration {
     #[must_use]
